ft_printf_hex.c now uses unsigned long type. Passed Tripuille tests !

+++ src/ft_printf.c
@@ -9,15 +9,15 @@ int	ft_printflag(char flag, va_list args)
 	if (flag == 's')
 		return ft_putstr(va_arg(args, char*));
 	if (flag == 'p')
-		return ft_putptr(va_arg(args, int));
+		return ft_putptr(va_arg(args, long));
 	if (flag == 'd' || flag == 'i')
 		return ft_putnbr(va_arg(args, int));
 	if (flag == 'u')
 		return ft_putnbr_uns(va_arg(args, unsigned));
 	if (flag == 'x')
-		return ft_puthex(va_arg(args, int), 0);
+		return ft_puthex(va_arg(args, unsigned), 0);
 	if (flag == 'X')
-		return ft_puthex(va_arg(args, int), 1);
+		return ft_puthex(va_arg(args, unsigned), 1);
 	if (flag == '%')
 		return ft_putchar('%');
 	return 0;
